861 adding methods from Executables and Fields to the new ElementUtils
