+++ processor/src/main/java/org/mapstruct/ap/internal/util/JavacElementUtilsDecorator.java
@@ -6,10 +6,16 @@
 package org.mapstruct.ap.internal.util;
 
 import javax.annotation.processing.ProcessingEnvironment;
+import javax.lang.model.element.TypeElement;
 
 public class JavacElementUtilsDecorator extends AbstractElementUtilsDecorator {
 
     JavacElementUtilsDecorator(ProcessingEnvironment processingEnv) {
         super( processingEnv );
+    }
+
+    @Override
+    protected TypeElement replaceTypeElementIfNecessary(TypeElement element) {
+        return element;
     }
 }
