--- processor/src/main/java/org/mapstruct/ap/internal/util/SpecificCompilerWorkarounds.java
@@ -1,41 +1,0 @@
-/*
- * Copyright MapStruct Authors.
- *
- * Licensed under the Apache License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
- */
-package org.mapstruct.ap.internal.util;
-
-import javax.lang.model.element.TypeElement;
-
-/**
- * Contains workarounds for various quirks in specific compilers.
- *
- * @author Sjaak Derksen
- * @author Andreas Gudian
- */
-public class SpecificCompilerWorkarounds {
-    private SpecificCompilerWorkarounds() {
-    }
-
-    /**
-     * When running during Eclipse Incremental Compilation, we might get a TypeElement that has an UnresolvedTypeBinding
-     * and which is not automatically resolved. In that case, getEnclosedElements returns an empty list. We take that as
-     * a hint to check if the TypeElement resolved by FQN might have any enclosed elements and, if so, return the
-     * resolved element.
-     *
-     * @param elementUtils element utils
-     * @param element the original element
-     * @return the element freshly resolved using the qualified name, if the original element did not return any
-     *         enclosed elements, whereas the resolved element does return enclosed elements.
-     */
-    public static TypeElement replaceTypeElementIfNecessary(ElementUtils elementUtils, TypeElement element) {
-        if ( element.getEnclosedElements().isEmpty() ) {
-            TypeElement resolvedByName = elementUtils.getTypeElement( element.getQualifiedName() );
-            if ( resolvedByName != null && !resolvedByName.getEnclosedElements().isEmpty() ) {
-                return resolvedByName;
-            }
-        }
-        return element;
-    }
-
-}
